Move test vector generator to test utils
This makes it easier for other projects to generate VDAF test vectors
with the same format.
